Preserve tile comments from ingested file
Record tile-specific comments in the ingested file and include them in
the output.  This way you don't lose all the comments from ingesting the
file and then outputting it again.  However, it does still lose comments
which are not specific to any one tile (for example, comments at the
very start of the file).

Could potentially do something with this like automatically adding or
updating a special "last edited on <date>" comment whenever a tile is
edited in nhtiledit.  Right now, though, it just preserves existing
comments and doesn't provide any capacity to add new ones.